Make sansshell service easier to follow.
This was confusing since the server def was stuffed into logging along with the Version var.

Put Version with version.go
Move generic server parts into server.go
